钱包自检完成

+++ BossCommonClass/Classes/SingletonClass/Account/BossKnight/Model/AccountModel/BalanceMoneyModel.h
@@ -5,8 +5,8 @@
 @property (nonatomic, strong) NSString * idField;
 @property (nonatomic, strong) NSString * balanceMoney;
 @property (nonatomic, strong) NSString * createdAt;
-@property (nonatomic, assign) NSInteger freeMoney;
-@property (nonatomic, assign) NSInteger frozenMoney;
+@property (nonatomic, assign) CGFloat freeMoney;
+@property (nonatomic, assign) CGFloat frozenMoney;
 @property (nonatomic, assign) NSInteger ownerType;
 @property (nonatomic, strong) NSString * updatedAt;
